% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammaplot.R
\name{gammaplot}
\alias{gammaplot}
\title{Visualize gamma clustering}
\usage{
gammaplot(gm, clusters)
}
\arguments{
\item{gm}{A gmatrix object}

\item{clusters}{A list of length(n_genes) labeling each gene with a cluster}
}
\value{
A UMAP plot color-coded by cluster membership
}
\description{
This function loads a file as a matrix. It assumes that the first column
contains the rownames and the subsequent columns are the sample identifiers.
Any rows with duplicated row names will be dropped with the first one being
kepted.
}
