% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gctest.R
\name{gctest}
\alias{gctest}
\title{Perform gamma clustering}
\usage{
gctest(n_initial, n_final, gm, method = "affinity")
}
\arguments{
\item{n_initial}{The lower bound number of clusters}

\item{n_final}{The upper bound number of clusters}

\item{gm}{A gmatrix object}

\item{method}{Use the "distance" or "affinity" matrix}
}
\value{
A plot of the mean confidence for each cluster number
}
\description{
This function loads a file as a matrix. It assumes that the first column
contains the rownames and the subsequent columns are the sample identifiers.
Any rows with duplicated row names will be dropped with the first one being
kepted.
}
