% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcfuncs.R
\name{gmatrix}
\alias{gmatrix}
\title{Load a Matrix}
\usage{
gmatrix(
  x,
  partial = FALSE,
  squared = TRUE,
  normalization = FALSE,
  permuted = 0
)
}
\arguments{
\item{x}{A matrix of fold changes with shape (n_genes, n_samples)}

\item{partial}{Replace pearson correlations with partial correlations}

\item{squared}{Square correlation values}

\item{normalization}{Perform quantile normalization on x}

\item{permuted}{Iterations of permutation testing for p values}
}
\value{
An object containing processed matrices
}
\description{
This function loads a file as a matrix. It assumes that the first column
contains the rownames and the subsequent columns are the sample identifiers.
Any rows with duplicated row names will be dropped with the first one being
kepted.
}
